Give completion rules for exact command higher precedence than rules for glob-pattern
With this commit, if we have completion rules like this:

    complete * ...
    complete cmd1 ...

then when user types:

    % cmd1 <tab>

the second rule will be used because it is an exact match. The first rule will
only be used when there is command that exactly matches, e.g.:

    % cmd2 <tab>

This allows the first rule to become a fallback or default completion rule.